Encrypts the api key in the ds config
The current implementation saves the API key in cleartext
which is a potential security issue.

This change does not use the standard http settings component
and instead creates a custom config which gives the user a choice
of two urls for the logzio endpoint and which encrypts the api
key in the secure json data blob. This encrypted key is then only
decrypted on the backend in the plugin proxy.

The plugin proxy uses the decrypted key when adding a header
to all requests to logz.io. This is configured in the plugin.json
file in the routes section.